pyproject.toml
[project]
name = "coffee-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "googlemaps>=4.10.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-google-genai>=3.0.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
